requirements.txt
Flask~=3.1.0
flask-cors~=5.0.1
Flask-Bcrypt~=1.0.1
python-dotenv~=1.0.1
redis~=5.2.1
Werkzeug~=3.1.3
Flask-JWT-Extended~=4.7.1
Flask-Migrate~=4.1.0